Teensy Update

+++ main/control/param.h
@@ -1,9 +1,9 @@
-#define encoderRes 334    //Encoder Resolution 344 pulse/rotation
-#define gearRatio 9.28    //Gear Ratio 1:9.28
-#define contPeriod 10000.0   //Control Period - 100Hz to microsecond  1/100sec = 0.01sec
+#define encoderRes 334            //Encoder Resolution 344 pulse/rotation
+#define gearRatio 9.28f           //Gear Ratio 1:9.28
+#define contPeriod 10000.0f       //Control Period - 100Hz to microsecond  1/100sec = 0.01sec
 
 //PID Parameters
 //Motor-1
-const double Kp1 = 0.000001;
-const double Ki1 = 0.000001;
-const double Kd1 = 0.0;
+const float Kp1 = 0.000001;
+const float Ki1 = 0.000001;
+const float Kd1 = 0.0;
